+++ src/lib/providers/network-guard/index.tsx
@@ -9,7 +9,6 @@ import {
   useAccountStore,
   useCodeStore,
   useContractStore,
-  useNetworkStore,
   usePublicProjectStore,
 } from "lib/providers/store";
 import { formatUserKey } from "lib/utils";
@@ -30,7 +29,6 @@ export const NetworkGuard = observer(({ children }: NetworkGuardProps) => {
   const { setAccountUserKey, isAccountUserKeyExist } = useAccountStore();
   const { setCodeUserKey, isCodeUserKeyExist } = useCodeStore();
   const { setContractUserKey, isContractUserKeyExist } = useContractStore();
-  const { setNetworkUserKey, isNetworkUserKeyExist } = useNetworkStore();
   const { setProjectUserKey, isProjectUserKeyExist } = usePublicProjectStore();
 
   useEffect(() => {
@@ -40,8 +38,6 @@ export const NetworkGuard = observer(({ children }: NetworkGuardProps) => {
       setCodeUserKey(userKey);
       setContractUserKey(userKey);
       setProjectUserKey(userKey);
-
-      setNetworkUserKey(DEFAULT_ADDRESS);
     }
   }, [
     isHydrated,
@@ -49,7 +45,6 @@ export const NetworkGuard = observer(({ children }: NetworkGuardProps) => {
     setAccountUserKey,
     setCodeUserKey,
     setContractUserKey,
-    setNetworkUserKey,
     setProjectUserKey,
   ]);
 
@@ -61,7 +56,6 @@ export const NetworkGuard = observer(({ children }: NetworkGuardProps) => {
     !isAccountUserKeyExist() ||
     !isCodeUserKeyExist() ||
     !isContractUserKeyExist() ||
-    !isNetworkUserKeyExist() ||
     !isProjectUserKeyExist()
   )
     return <LoadingOverlay />;